Refactor some internal datastructures
Make board an object with position keys (e.g. "i8") and tile
number values instead of an array of arrays of tile number
values.  As the board is generally sparse (especially early
in the game), this leads to a shorter representation, and it
reduces the corner cases when checking the edges of the board.

Make Game hold an array of states, change state to a
function that returns the last/current state.

Fix bug in game/check where the json returned was not an
array.

Don't have states retain knowledge of previous states.