Huge, huge changes
Most of the changes relate to now handling multiple players and
games on the server.  This includes registering new players,
logging in existing players, starting new games, and joining
existing games.  Persistence is currently done with a json
file tree, but that's done in a separate module so it can
easily be swapped out.