Remove nodejs-related code
The nodejs server doesn't appear to work in newer versions of nodejs,
and I have no desire to fix it.  As the go server works better,
it should be used instead.

There is currently no replacement for the command line client.
Considering how much friendlier a web browser is, it's
unlikely anyone would prefer the command line client.

As node is no longer used, switch from a Cakefile to a Rakefile
for running the tests.  Ruby is required anyway since the
web server integration tests are written in it.