Use long polling for moves for faster response time
If long polling fails, try again in 1 minute.  If no results
are received when long polling, try another long poll after
10 minutes.  Don't show the loading icon when long polling.

This is a simplistic implementation that only works if the webserver
is a single process.  To expand to multiple processes/servers,
it should use LISTEN/NOTIFY via PostgreSQL to get notified of
moves, or a real message queue.  Currently, the demo instance
only runs on a single dyno on Heroku, so this is safe for that
purpose.

I'm not sure whether this leakes memory via channels or not,
it probably does.