Use MessageBus for real-time notifications of moves
This uses message_bus only to notify clients that a move has been
made.  The json the clients use is customized per client, so after
receiving the notification from message bus, they check with the
server to get the current state.

Remove the spinner, now that we are long polling/streaming using
message bus, it doesn't make sense to spin during requests.

This currently only uses the message_bus memory backend. It may in
the future switch to using the message_bus PostgreSQL backend, or
make the backend configurable.