Pack finished games
When a game is finished, this removes the game states from the
game_states table, and moves them to a game_states[] column in
the games table.  In an environment where the number of allowed
rows is limited, this allows more finished games to be stored.

Use the pg_array and pg_row Sequel extensions to handle the array
of composite types.  Add 3 prepared statements:

* PackFinishedGame: Copy the game_states from the game_states table
  to a single column in the games table.
* FinshedGameStates: Get the packed finished game states for a game.
* DeleteFinishedGameStates: Delete game states for a game.

When persisting a move, if the game has ended, pack the finished game
and then delete the game states.

When looking up a finished game, there will no longer be a row in the
game_states table, so look at the finished game states in that case.

Update the stats to operate on the finished game states, since it
already was limited to looking at finished games.