Switch game_players table to a player_ids column in the games table
This also reduces the number of rows.  In some cases it is simpler,
since an array of player ids has an implicit order and there isn't
a need for a position column.

Change the code that needs both a game ID and player emails to get
player ids in one query and emails in a separate query. This results
in simpler queries, and the performance difference doesn't matter.

Call Player#active_and_finished_games to get the active and finished
games instead of separate methods.  This way a single query can be
used to get the game information.